backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
redis
routeros-api 